+++ src/app/(data)/items/mod-items.ts
@@ -372,7 +372,7 @@ export const modItems: ModItem[] = [
     imagePath: '/items/mods/cyclone.png',
     id: '1bfadq',
     dlc: 'dlc2',
-    description: `Maintaining an overhead Charged Melee forms a 5m sand cyclone, dealing 55 - 168 damage per second and pulls enemies towards the wielder. Each Cyclone hit applies EXPOSED for 3s. Max 15s.`,
+    description: `Maintaining an overhead Charged Melee forms a 5m sand cyclone, dealing 35 damage per second and pulls enemies towards the wielder. Each Cyclone hit applies EXPOSED for 3s. Max 15s.`,
     wikiLinks: [`https://remnant.wiki/Cyclone`],
     linkedItems: {
       weapon: {
@@ -1081,9 +1081,10 @@ export const modItems: ModItem[] = [
     id: '2JimAt',
     dlc: 'dlc2',
     tags: ['Status Effect'],
-    description: 'Charged Melee empowers the staff for 10s, causing Melee Attacks to launch an energy bolt that seeks enemies within 30m, dealing 50 - 150 SHOCK Damage and grants 1 stack of POWER STONE. Max 10 stacks. \n' +
-    '\n' +
-    'While empowered, Charged Melee calls a 6m shockwave dealing 75 - 225 SHOCK Damage and ends POWER STONE. For each POWER STONE stack built when ended, release 1 energy bolt that seeks a random enemy and applies OVERLOADED to 10s.',
+    description:
+      'Charged Melee empowers the staff for 10s, causing Melee Attacks to launch an energy bolt that seeks enemies within 30m, dealing 50 - 150 SHOCK Damage and grants 1 stack of POWER STONE. Max 10 stacks. \n' +
+      '\n' +
+      'While empowered, Charged Melee calls a 6m shockwave dealing 75 - 225 SHOCK Damage and ends POWER STONE. For each POWER STONE stack built when ended, release 1 energy bolt that seeks a random enemy and applies OVERLOADED to 10s.',
     externalTokens: [`Amplitude`],
     wikiLinks: [`https://remnant.wiki/Power_Stone`],
     linkedItems: {