+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-05-22
+
+### Changed
+
+- Favorite Build button will now show up whether the user is signed in or not - previously would only show if you are signed in. If an unauthenticated user clicks the button, they will be prompted to sign-in to favorite the build.
+
+### Fixed
+
+- Updated Mirage weapon description and damage values.
+- Updated Cyclone mod description.
+
 ## 2024-05-21
 
 ### Added